+++ engine/ratings/__init__.py
@@ -1,0 +1,25 @@
+from engine.ratings.sector_rating import (
+    ADVANTAGE_BALANCED,
+    ADVANTAGE_NOT_COMPARABLE,
+    ADVANTAGE_OPPONENT,
+    ADVANTAGE_OURS,
+    SOURCE_HATTRICK_DECIMAL,
+    SOURCE_HT_COACH_INTERNAL,
+    SectorComparison,
+    build_sector_comparisons,
+    detect_rating_scale,
+    format_rating_value,
+)
+
+__all__ = [
+    "ADVANTAGE_BALANCED",
+    "ADVANTAGE_NOT_COMPARABLE",
+    "ADVANTAGE_OPPONENT",
+    "ADVANTAGE_OURS",
+    "SOURCE_HATTRICK_DECIMAL",
+    "SOURCE_HT_COACH_INTERNAL",
+    "SectorComparison",
+    "build_sector_comparisons",
+    "detect_rating_scale",
+    "format_rating_value",
+]